REFACTOR - change test directories of test_crawlers
